+++ src/MadPrintout.h
@@ -16,7 +16,7 @@ public:
     MadPrintout(const wxString& title = _("MadEdit-Mod Printout"));
     virtual ~MadPrintout();
 
-    virtual void GetPageInfo(int *minPage, int *maxPage, int *pageFrom, int *pageTo);
+    virtual void GetPageInfo(size_t *minPage, size_t *maxPage, size_t *pageFrom, size_t *pageTo);
     virtual bool HasPage(int page);
     virtual bool OnPrintPage(int page);
